Submitting modifications to selectionSort.cpp and sort.cpp

+++ C++/sorting/selectionSort.cpp
@@ -3,9 +3,9 @@
  * OuterIndex for Swap and InnerIndex to find minIndex.
  */
 void SortClass::selectionSortIterative(){
-    for(size_t i = 0; i < sz - 1; i++){
+    for(int i = 0; i < sz - 1; i++){
         int min = i;
-        for(size_t j = i + 1; j < sz; j++){
+        for(int j = i + 1; j < sz; j++){
             if( arr[j] < arr[min] ){
                 min = j;
             }
@@ -14,9 +14,9 @@ void SortClass::selectionSortIterative(){
     }
 }
 
-size_t SortClass::findMinIndex(size_t startPos){
-    size_t min = startPos;
-    for(size_t j = startPos; j < sz; j++){
+int SortClass::findMinIndex(int startPos){
+    int min = startPos;
+    for(int j = startPos; j < sz; j++){
         if( arr[j] < arr[min]){
             min = j;
         }
@@ -24,11 +24,11 @@ size_t SortClass::findMinIndex(size_t startPos){
     return min;
 }
 
-void SortClass::selectionSortRecursive(size_t startPos){
+void SortClass::selectionSortRecursive(int startPos){
     if(startPos == sz){
         return; 
     }
-    size_t minPos = findMinIndex(startPos);
+    int minPos = findMinIndex(startPos);
     int temp = arr[startPos];
     arr[startPos] = arr[minPos];
     arr[minPos] = temp;
